+++ docs/changelog.md
@@ -2,6 +2,9 @@
 
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html) and [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) format.
 
+## [0.7.1] - 2024-xx-xx
+### Added
+- added `portable` flag to pipestat summarize to create a shareable version of the html report.
 
 ## [0.7.0] - 2024-01-17
 ### Added